[Decomp][Player] Cohesion split: extract battleground methods to PlayerBattleGround.cpp
